do 3.1 and 3.2

+++ c5/vector.h
@@ -14,7 +14,7 @@
 typedef struct
 {
         unsigned int length;
-        long double* element;
+        double* element;
 } Vector;
